refactor: reorganize chips
