+++ package.json
@@ -29,6 +29,7 @@
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
         "react-tagcloud": "^2.3.3",
+        "use-immer": "^0.10.0",
         "video-privacy": "^1.0.2"
     },
     "devDependencies": {
