--- src/components/eln/SubdiscChip.js
@@ -1,5 +1,0 @@
-function SubdiscChip({ subdisc }) {
-    return <button className={"lbe__filterbutton "}>{subdisc}</button>;
-}
-
-export default SubdiscChip;
